Redoing restructure to avoid downstream merge conflicts
