+++ environment.yml
@@ -3,7 +3,7 @@ channels:
   - defaults
   - conda-forge
 dependencies:
-  - python >=3.9, <4.0
+  - python >=3.9, <=3.10
   - jupyter
   - ipykernel
   - pandas
