--- tests/test_data/test.json
@@ -1,40 +1,0 @@
-{
-    "title": "A Science-Based Protocol for Testing indexing.py (But Seriously, Don't Get Rickrolled)",
-    "ep_num": "1337",
-    "link": "https://www.youtube.com/watch?v=dQw4w9WgXcQ",
-    "chunks": [
-        {
-            "timestamp": 0,
-            "text": "Intriguing hypothesis detected. Commence analysis!"
-        },
-        {
-            "timestamp": 186,
-            "text": "Wait, is this song stuck in my head now?  (Focus, science!)"
-        },
-        {
-            "timestamp": 436,
-            "text": "Data point confirmed. Indexing appears functional."
-        },
-        {
-            "timestamp": 685,
-            "text": "But seriously, why are bananas so bendy?"
-        },
-        {
-            "timestamp": 878,
-            "text": "Indexing complete. Prepare for celebratory cat videos!"
-        },
-        {
-            "timestamp": 1080,
-            "text": "Wait, where were we? Oh right, science stuff."
-        },
-        {
-            "timestamp": 1478,
-            "text": "Fascinating! This warrants further investigation with... more cat videos?"
-        },
-        {
-            "timestamp": 1769,
-            "text": "Conclusion: indexing.py seems to work. Now go take a break, scientist!"
-        }
-    ],
-    "episode_summary": "A (mostly) successful exploration into the scientific marvel of indexing.py (with some cat-related tangents)."
-}
