+++ tldhuber/tests/test_transcripts_scraper.py
@@ -13,9 +13,12 @@ import os
 import json
 import pandas as pd
 
-from tldhuber.utils.transcripts_scraper import (get_channel_upload_playlist_id_by_username,
-                                      get_channel_upload_playlist_id_by_channelid,
-                                      get_playlist_items, save_video_information_to_json)
+from tldhuber.utils.transcripts_scraper import (
+    get_channel_upload_playlist_id_by_username,
+    get_channel_upload_playlist_id_by_channelid,
+    get_playlist_items,
+    save_video_information_to_json,
+)
 
 
 class TestTranscriptsScraper(unittest.TestCase):
@@ -27,7 +30,7 @@ class TestTranscriptsScraper(unittest.TestCase):
     information to JSON.
     """
 
-    @patch('requests.get')
+    @patch("requests.get")
     def test_get_channel_upload_playlist_id_by_username(self, mock_get):
         """
         Test fetching the upload playlist ID using a YouTube channel username.
@@ -36,9 +39,7 @@ class TestTranscriptsScraper(unittest.TestCase):
             "items": [
                 {
                     "contentDetails": {
-                        "relatedPlaylists": {
-                            "uploads": "UPLOAD_PLAYLIST_ID"
-                        }
+                        "relatedPlaylists": {"uploads": "UPLOAD_PLAYLIST_ID"}
                     }
                 }
             ]
@@ -48,7 +49,7 @@ class TestTranscriptsScraper(unittest.TestCase):
         result = get_channel_upload_playlist_id_by_username(api_key, username)
         self.assertEqual(result, "UPLOAD_PLAYLIST_ID")
 
-    @patch('requests.get')
+    @patch("requests.get")
     def test_get_channel_upload_playlist_id_by_channelid(self, mock_get):
         """
         Test fetching the upload playlist ID using a YouTube channel ID.
@@ -57,9 +58,7 @@ class TestTranscriptsScraper(unittest.TestCase):
             "items": [
                 {
                     "contentDetails": {
-                        "relatedPlaylists": {
-                            "uploads": "UPLOAD_PLAYLIST_ID"
-                        }
+                        "relatedPlaylists": {"uploads": "UPLOAD_PLAYLIST_ID"}
                     }
                 }
             ]
@@ -69,40 +68,56 @@ class TestTranscriptsScraper(unittest.TestCase):
         result = get_channel_upload_playlist_id_by_channelid(api_key, channel_id)
         self.assertEqual(result, "UPLOAD_PLAYLIST_ID")
 
-    @patch('requests.get')
+    @patch("requests.get")
     def test_get_playlist_items(self, mock_get):
         """
         Test retrieving all items from a YouTube playlist.
         """
         mock_get.return_value.json.side_effect = [
             {"items": ["video1", "video2"], "nextPageToken": "token"},
-            {"items": ["video3"]}
+            {"items": ["video3"]},
         ]
         api_key = "dummy_api_key"
         playlist_id = "dummy_playlist_id"
         result = get_playlist_items(api_key, playlist_id)
         self.assertEqual(result, ["video1", "video2", "video3"])
 
+    @patch('requests.get')
+    def test_get_channel_upload_playlist_id_by_username_no_items(self, mock_get):
+        """
+        Test upload playlist ID using a YouTube channel username when no items are returned.
+        This is an edge case where the given username does not exist or there is an API issue.
+        """
+        mock_get.return_value.json.return_value = {
+            "items": []
+        }
+        api_key = "dummy_api_key"
+        username = "dummy_username"
+        result = get_channel_upload_playlist_id_by_username(api_key, username)
+        self.assertIsNone(result)
+
     def test_save_video_information_to_json(self):
         """
         Test saving video information to a JSON file.
         """
-        video_information = pd.DataFrame({
-            'Video Title': ['Title 1'],
-            'Video URL': ['http://example.com'],
-            'Transcript': ['This is a transcript.']
-        })
+        video_information = pd.DataFrame(
+            {
+                "Video Title": ["Title 1"],
+                "Video URL": ["http://example.com"],
+                "Transcript": ["This is a transcript."],
+            }
+        )
         file_path = "test_output.json"
         save_video_information_to_json(video_information, file_path)
 
-        with open(file_path, 'r', encoding='utf-8') as file:
+        with open(file_path, "r", encoding="utf-8") as file:
             content = json.load(file)  # Deserialize JSON content to a Python object
-            self.assertEqual(content['Video URL'], 'http://example.com')
-            self.assertEqual(content['Video Title'], 'Title 1')
-            self.assertEqual(content['Transcript'], 'This is a transcript.')
+            self.assertEqual(content["Video URL"], "http://example.com")
+            self.assertEqual(content["Video Title"], "Title 1")
+            self.assertEqual(content["Transcript"], "This is a transcript.")
 
         os.remove(file_path)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     unittest.main()
